requirements.txt
pandas==1.1.3
streamlit==0.81.1
matplotlib==3.3.2
pymongo==3.11.3
altair==4.1.0
numpy==1.19.2
dnspython==2.1.0
Pillow==8.2.0
